requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.64.0
dataclasses>=0.6
typing-extensions>=4.0.0

# Visualization & Metrics
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Logging & Development
tensorboard>=2.8.0
wandb>=0.13.0

